refactor: move components in specific folder
