+++ website/src/components/Legend.svelte
